fix: update coming soon for store
