+++ src/pages/main/Library.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from "react-i18next";
 import { Grid, Box } from "@mui/material";
 
 import LibraryModeBtn from "../../components/library/LibraryMode";
-import Multichainbtn from "../../components/store/Multichainbtn";
+import MultiChainButton from "../../components/store/MultiChainButton";
 import LibraryShow from "../../components/library/LibraryShow";
 import StoreComingGameItems from "../../components/store/StoreComingGameItems";
 
@@ -17,7 +17,7 @@ const Library = () => {
       <Box className={"fs-60-bold white"}>{t("hom-3_library")}</Box>
       <Grid item xs={12} container display={"flex"} marginTop={"48px"} justifyContent={"space-between"}>
         <LibraryModeBtn status={status} setStatus={setStatus} />
-        {false && <Multichainbtn />}
+        {false && <MultiChainButton />}
       </Grid>
       <Grid item xs={12} marginTop={"32px"}>
         <LibraryShow status={status} />
